refactor: extract api calls and support constant
